Implemented getAnimeList and updateFavorite in the AnimeRepositoryImpl. Updated AnimeDao to return Int from updateFavorite. Updated score in AnimeDto to be a Double? instead of Int?. Renamed QueryParametersMapper to QueryParametersDomainToDataMapper to be consistent with the other mappers. Added AnimeDataToDomainMapper, AnimeDomainToDataMapper and StringToLocalDateMapper.

+++ app/src/main/java/app/vercel/lcsanimelist/data/network/dto/AnimeDto.kt
@@ -8,7 +8,7 @@ data class AnimeDto(
     @SerializedName("title") val title: String,
     @SerializedName("episodes") val episodes: Int?,
     @SerializedName("aired") val aired: AiredDateDto,
-    @SerializedName("score") val score: Int?,
+    @SerializedName("score") val score: Double?,
     @SerializedName("synopsis") val synopsis: String?,
     @SerializedName("genres") val genres: List<GenreDto>
 )
